Move tests to parent namespace for now
